// apps/admin-web/src/api/model/updateConfigItemDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * I 54KB API
 * NestJS 企业级后端服务 API，基于 Prisma + PostgreSQL + Redis + Socket.io，提供用户认证、审计日志、幂等性控制等核心能力
 * OpenAPI spec version: 1.0
 */
import type { UpdateConfigItemDtoValue } from "./updateConfigItemDtoValue";
import type { UpdateConfigItemDtoValueType } from "./updateConfigItemDtoValueType";
import type { UpdateConfigItemDtoJsonSchema } from "./updateConfigItemDtoJsonSchema";

export interface UpdateConfigItemDto {
  value?: UpdateConfigItemDtoValue;
  valueType?: UpdateConfigItemDtoValueType;
  /** @maxLength 500 */
  description?: string;
  isEncrypted?: boolean;
  isPublic?: boolean;
  isEnabled?: boolean;
  jsonSchema?: UpdateConfigItemDtoJsonSchema;
}